Move transaction package to safe package; rename transactions fragment
