+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -8,7 +8,7 @@
         android:title="@string/assets"
         app:showAsAction="always" />
     <item
-        android:id="@+id/transactionsFragment"
+        android:id="@+id/transactionListFragment"
         android:enabled="true"
         android:icon="@drawable/ic_transactions_24dp"
         android:title="@string/transactions"
